feat: validity check
Check whether a figure can move to the desired location, thus does not enter the housing space of an other player. If the destination would be the housing space of an other player, then the destination's identifier will be increased by the amount of steps that the figure would be inside the housing space.