Finished Assignment Sheet 7 (#10)
* Project Initialization

* chore: added Internet permissions

* feat: added input, output and button to the layout

* chore: ignored env class

* feat: implemented API handler

This handler fetches data from an api and stores the response locally. If the data is requests a second time it will be loaded from the local storage.

* feat: fetching for specific currencies

Fetching for a an array of currencies with  a size of 2. Checking if the locally stored data includes those currencies. Fetches rates if they are not stored. Returning the rates object.

* feat: added rate fields - layout

* feat: added MainActivity functionalities

* feat: added currency enum class

* feat: filled spinner with dynamic currency options

* feat: implemented DB handler for SQL Lite

Added a DBHandler that initializes a SQL Lite database.

* refactor: externalized date-formatter

* feat: added function that inserts data - API

* feat: added function that reads from the database

Reads stored entries from the database. If nothing is stored the currency factors are fetched from the api.

* feat: added check if stored factors match request

* refactor: renamed API functions

* refactor: externalized local storage functionality

* feat: added dynamic storage functionality

Now storing new fetch request either locally (file) or remote (db) depending on the configuration.

* feat: added dynamic function for getting factors

When calling this method the stored data will be returned, if no data has been stored, then it will be fetched and stored. Note, this method will either attempt to read and store data locally or remote depending on the configuration.

* refactor: modified visibility modifiers - API

The API handler now has now one public function that differs between local and remote storage, based on the configuration.

* feat: added configuration variable - MainActivity

* refactor: removed redundant modifiers

* docs: added kotlin-doc for getDBData

* docs: added kotlin-doc for storeDB

* docs: added kotlin-doc for storeLocal

* docs: added kotlin-doc for getLocalData

* docs: added kotlin-doc for getData